规格参数CRUD

+++ mingrui-shop-parent/mingrui-shop-service/mingrui-shop-service-xxx/src/main/java/com/mr/shop/service/impl/SpecificationServiceImpl.java
@@ -3,12 +3,16 @@ package com.mr.shop.service.impl;
 import com.baidu.shop.base.BaseApiService;
 import com.baidu.shop.base.Result;
 import com.baidu.shop.dto.SpecGroupDTO;
+import com.baidu.shop.dto.SpecParamDTO;
 import com.baidu.shop.entity.SpecGroupEntity;
+import com.baidu.shop.entity.SpecParamEntity;
 import com.baidu.shop.service.SpecificationService;
 import com.baidu.shop.utils.BeanUtil;
 import com.baidu.shop.utils.ObjectUtil;
 import com.google.gson.JsonObject;
 import com.mr.shop.mapper.SpecGroupMapper;
+import com.mr.shop.mapper.SpecParamMapper;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RestController;
 import tk.mybatis.mapper.entity.Example;
 
@@ -28,6 +32,9 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
     @Resource
     private SpecGroupMapper specGroupMapper;
 
+    @Resource
+    private SpecParamMapper specParamMapper;
+
     @Override
     public Result<List<SpecGroupEntity>> list(SpecGroupDTO specGroupDTO) {
 
@@ -39,6 +46,7 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
         return this.setResultSuccess(list);
     }
 
+    @Transactional
     @Override
     public Result<JsonObject> add(SpecGroupDTO specGroupDTO) {
 
@@ -46,6 +54,7 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
         return this.setResultSuccess();
     }
 
+    @Transactional
     @Override
     public Result<JsonObject> edit(SpecGroupDTO specGroupDTO) {
 
@@ -53,12 +62,51 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
         return this.setResultSuccess();
     }
 
+    @Transactional
     @Override
     public Result<JsonObject> delete(Integer id) {
+
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId", id);
+        List<SpecParamEntity> list = specParamMapper.selectByExample(example);
+
+        if(list.size() > 0) return this.setResultError("此规格被绑定参数不能被删除");
 
         specGroupMapper.deleteByPrimaryKey(id);
         return this.setResultSuccess();
     }
 
+    @Override
+    public Result<List<SpecParamEntity>> listParam(SpecParamDTO specParamDTO) {
 
+        if(ObjectUtil.isNull(specParamDTO.getGroupId())) return this.setResultError("id不能为空");
+
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId", specParamDTO.getGroupId());
+        List<SpecParamEntity> list = specParamMapper.selectByExample(example);
+
+        return this.setResultSuccess(list);
+    }
+
+    @Transactional
+    @Override
+    public Result<JsonObject> addParam(SpecParamDTO specParamDTO) {
+
+        specParamMapper.insertSelective(BeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    @Override
+    public Result<JsonObject> editParam(SpecParamDTO specParamDTO) {
+
+        specParamMapper.updateByPrimaryKeySelective(BeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    @Override
+    public Result<JsonObject> deleteParam(Integer id) {
+
+        specParamMapper.deleteByPrimaryKey(id);
+        return this.setResultSuccess();
+    }
 }
